+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 - [PostgreSQL Dialect] Add basic Array literal support (#5997 by @griffio)
 - [PostgreSQL Dialect] Add basic LTREE support (#5880 by @yesitskev @griffio)
 - [MySQL Dialect] Add support for INET functions (#5072 by @mcxinyu)
+- [PostgreSQL Dialect] Add support for ALTER INDEX (#6224 by @griffio)
 - [SQLite Dialect] Add support for SQLite 3.44 aggregate functions DISTINCT, ORDER BY and FILTER (#6236 by @griffio)
 - [SQLite Dialect] Add support for SQLite 3.37 STRICT table (#6230 by @griffio)
 - [Gradle Plugin] Add support for excluding columns from generated models with `codegenExcludedColumns` (#6243 by @sokolikp)
+- [Compiler] Add `allTableNames` function to schema (#6245 by @edenman)
 
 ### Changed
 - [PostgreSQL Dialect] Change arrayIntermediateType visibility to public (#5835 by @griffio)
@@ -24,6 +26,7 @@
 - [Intellij Plugin] IDE plugin can perform code completions for all dialects (#6210 by @griffio)
 - [Gradle Plugin] Fix circular dependency error running verify database task (#6221 by @griffio)
 - [Compiler] Fix optimistic lock for multirow update (#6240 by @griffio)
+- [Intellij Plugin] Fix deprecations causing crash in IDEA 2026.2 (#6247 by @griffio)
 
 ## [2.3.2] - 2026-03-16
 [2.3.2]: https://github.com/sqldelight/sqldelight/releases/tag/2.3.2